+++ src/examples/java/com/nexmo/client/messaging/examples/SendTextMessage.java
@@ -19,22 +19,19 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.messaging.sdk.examples.signed;
+package com.nexmo.client.messaging.examples;
 
 
-import com.nexmo.messaging.sdk.NexmoSmsClient;
-import com.nexmo.messaging.sdk.NexmoSmsClientSignedRequests;
-import com.nexmo.messaging.sdk.SmsSubmissionResult;
-import com.nexmo.messaging.sdk.messages.TextMessage;
+import com.nexmo.client.messaging.NexmoSmsClient;
+import com.nexmo.client.messaging.SmsSubmissionResult;
+import com.nexmo.client.messaging.messages.TextMessage;
 
 /**
  * An example of using the Nexmo SMS API to submit a simple text message.
- * This example uses a {@link com.nexmo.messaging.sdk.NexmoSmsClientSignedRequests} client
- * So that the REST requests are signed using the secret key.
  *
  * @author  Paul Cook
  */
-public class SendSignedTextMessage {
+public class SendTextMessage {
 
     public static final String API_KEY = "your-api-key-goes-here";
     public static final String API_SECRET = "your-api-secret-goes-here";
@@ -49,7 +46,7 @@ public class SendSignedTextMessage {
 
         NexmoSmsClient client = null;
         try {
-            client = new NexmoSmsClientSignedRequests(API_KEY, API_SECRET);
+            client = new NexmoSmsClient(API_KEY, API_SECRET);
         } catch (Exception e) {
             System.err.println("Failed to instantiate a Nexmo Client");
             e.printStackTrace();
