+++ src/main/java/com/nexmo/client/messaging/SmsSubmissionResult.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.messaging.sdk;
+package com.nexmo.client.messaging;
 
 
 import java.math.BigDecimal;
