+++ src/main/java/com/nexmo/client/messaging/messages/parameters/ValidityPeriod.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.messaging.sdk.messages.parameters;
+package com.nexmo.client.messaging.messages.parameters;
 
 
 import java.util.Calendar;
